+++ augmentLTI.h
@@ -7,16 +7,16 @@
 /// @param Ad Matrix of the discrete LTI system
 /// @param Bd Matrix of the discrete LTI system
 /// @return Matrix At of the augmented discrete system
-Eigen::MatrixXf augmentA(Eigen::MatrixXf &Ad, Eigen::MatrixXf &Bd);
+Eigen::MatrixXd augmentA(Eigen::MatrixXd &Ad, Eigen::MatrixXd &Bd);
 
 /// @brief Creates the Bt matrix of the augmented system
 /// @param Bd Matrix of the discrete LTI system
 /// @return Matrix Bt of the augmented discrete system
-Eigen::MatrixXf augmentB(Eigen::MatrixXf &Bd);
+Eigen::MatrixXd augmentB(Eigen::MatrixXd &Bd);
 
 /// @brief Creates the Ct output matrix of the augmented system
 /// @param Cd Matrix of the discrete LTI system
 /// @return Matrix Ct of the augmented discrete system
-Eigen::MatrixXf augmentC(Eigen::MatrixXf &Cd);
+Eigen::MatrixXd augmentC(Eigen::MatrixXd &Cd);
 
 #endif
