+++ discretizeLTI.h
@@ -7,12 +7,12 @@
 /// @param A State space matrix of the continuous system (LTI)
 /// @param h integration interval
 /// @return Ad matrix of the disctete LTI
-Eigen::MatrixXf discretizeA(Eigen::MatrixXf &A, float h);
+Eigen::MatrixXd discretizeA(Eigen::MatrixXd &A, double h);
 
 /// @brief Build the discretized Bd matrix from the system state space B matrix
 /// @param B State space matrix B of the continuous system (LTI)
 /// @param h integration interval
 /// @return Bd matrix of the discrete LTI
-Eigen::MatrixXf discretizeB(Eigen::MatrixXf &B, float h);
+Eigen::MatrixXd discretizeB(Eigen::MatrixXd &B, double h);
 
 #endif
